+++ database/migrations/2025_05_01_005840_add_votes_cascade_delete_handling.php
@@ -7,44 +7,48 @@ class AddVotesCascadeDeleteHandling extends Migration
 {
     public function up()
     {
-        // Xóa trigger cũ nếu tồn tại
-        DB::unprepared('DROP TRIGGER IF EXISTS tr_votes_on_question_delete');
+        if (DB::getDriverName() === 'sqlsrv') {
+            // Xóa trigger cũ nếu tồn tại (nếu có dùng trigger)
+            DB::unprepared('DROP TRIGGER IF EXISTS tr_votes_on_question_delete');
 
-        // Tạo stored procedure xử lý xóa question
-        DB::unprepared('
-            IF NOT EXISTS (SELECT * FROM sys.objects WHERE type = \'P\' AND name = \'sp_delete_question_with_votes\')
-            BEGIN
-                EXEC(\'
-                    CREATE PROCEDURE sp_delete_question_with_votes(@question_id INT)
-                    AS
-                    BEGIN
-                        BEGIN TRANSACTION;
-                        DELETE FROM votes WHERE question_id = @question_id;
-                        DELETE FROM comments WHERE question_id = @question_id;
-                        DELETE FROM questions WHERE id = @question_id;
-                        COMMIT;
-                    END
-                \')
-            END
-            ELSE
-            BEGIN
-                EXEC(\'
-                    ALTER PROCEDURE sp_delete_question_with_votes(@question_id INT)
-                    AS
-                    BEGIN
-                        BEGIN TRANSACTION;
-                        DELETE FROM votes WHERE question_id = @question_id;
-                        DELETE FROM comments WHERE question_id = @question_id;
-                        DELETE FROM questions WHERE id = @question_id;
-                        COMMIT;
-                    END
-                \')
-            END
-        ');
+            // Tạo hoặc sửa stored procedure
+            DB::unprepared('
+                IF NOT EXISTS (SELECT * FROM sys.objects WHERE type = \'P\' AND name = \'sp_delete_question_with_votes\')
+                BEGIN
+                    EXEC(\'
+                        CREATE PROCEDURE sp_delete_question_with_votes(@question_id INT)
+                        AS
+                        BEGIN
+                            BEGIN TRANSACTION;
+                            DELETE FROM votes WHERE question_id = @question_id;
+                            DELETE FROM comments WHERE question_id = @question_id;
+                            DELETE FROM questions WHERE id = @question_id;
+                            COMMIT;
+                        END
+                    \')
+                END
+                ELSE
+                BEGIN
+                    EXEC(\'
+                        ALTER PROCEDURE sp_delete_question_with_votes(@question_id INT)
+                        AS
+                        BEGIN
+                            BEGIN TRANSACTION;
+                            DELETE FROM votes WHERE question_id = @question_id;
+                            DELETE FROM comments WHERE question_id = @question_id;
+                            DELETE FROM questions WHERE id = @question_id;
+                            COMMIT;
+                        END
+                    \')
+                END
+            ');
+        }
     }
 
     public function down()
     {
-        DB::unprepared('DROP PROCEDURE IF EXISTS sp_delete_question_with_votes');
+        if (DB::getDriverName() === 'sqlsrv') {
+            DB::unprepared('DROP PROCEDURE IF EXISTS sp_delete_question_with_votes');
+        }
     }
 }
